+++ src/Umbraco.Abstractions/Models/EntityExtensions.cs
