+++ src/Umbraco.Abstractions/Models/MemberTypePropertyProfileAccess.cs
@@ -3,7 +3,7 @@
     /// <summary>
     /// Used to track the property types that are visible/editable on member profiles
     /// </summary>
-    internal class MemberTypePropertyProfileAccess
+    public class MemberTypePropertyProfileAccess
     {
         public MemberTypePropertyProfileAccess(bool isVisible, bool isEditable, bool isSenstive)
         {
