+++ src/Umbraco.Core/Umbraco.Core.csproj
@@ -213,6 +213,7 @@
     <Compile Include="Migrations\Upgrade\V_8_0_1\ChangeNuCacheJsonFormat.cs" />
     <Compile Include="Migrations\Upgrade\V_8_1_0\ConvertTinyMceAndGridMediaUrlsToLocalLink.cs" />
     <Compile Include="Models\ContentBaseExtensions.cs" />
+    <Compile Include="Models\GridValue.cs" />
     <Compile Include="Models\IContentType.cs" />
     <Compile Include="Models\IContentTypeBase.cs" />
     <Compile Include="Models\IContentTypeComposition.cs" />
@@ -225,12 +226,10 @@
     <Compile Include="Models\IMemberType.cs" />
     <Compile Include="Models\Language.cs" />
     <Compile Include="Models\Macro.cs" />
-    <Compile Include="Models\MacroProperty.cs" />
     <Compile Include="Models\Membership\UserGroup.cs" />
     <Compile Include="Models\Membership\UserGroupExtensions.cs" />
     <Compile Include="Models\ObjectTypes.cs" />
     <Compile Include="Models\Packaging\PackageDefinition.cs" />
-    <Compile Include="Models\PublicAccessEntry.cs" />
     <Compile Include="Migrations\Upgrade\V_8_0_0\FixLanguageIsoCodeLength.cs" />
     <Compile Include="Models\CultureImpact.cs" />
     <Compile Include="Models\PublishedContent\IndexedArrayItem.cs" />
@@ -524,25 +523,17 @@
     <Compile Include="Models\ContentTypeBase.cs" />
     <Compile Include="Models\ContentTypeCompositionBase.cs" />
     <Compile Include="Models\DataType.cs" />
-    <Compile Include="Models\DictionaryItem.cs" />
-    <Compile Include="Models\DictionaryItemExtensions.cs" />
-    <Compile Include="Models\DictionaryTranslation.cs" />
-    <Compile Include="Models\EntityExtensions.cs" />
-    <Compile Include="Models\GridValue.cs" />
     <Compile Include="Models\Identity\BackOfficeIdentityUser.cs" />
     <Compile Include="Models\Media.cs" />
     <Compile Include="Models\MediaExtensions.cs" />
     <Compile Include="Models\MediaType.cs" />
     <Compile Include="Models\Member.cs" />
-    <Compile Include="Models\MemberGroup.cs" />
     <Compile Include="Models\Membership\MembershipUserExtensions.cs" />
     <Compile Include="Models\Membership\UmbracoMembershipMember.cs" />
     <Compile Include="Models\Membership\UmbracoMembershipUser.cs" />
     <Compile Include="Models\Membership\User.cs" />
     <Compile Include="Models\MemberType.cs" />
-    <Compile Include="Models\MemberTypePropertyProfileAccess.cs" />
     <Compile Include="Models\Packaging\InstallationSummary.cs" />
-    <Compile Include="Models\PagedResultOfT.cs" />
     <Compile Include="Models\Property.cs" />
     <Compile Include="Models\PropertyCollection.cs" />
     <Compile Include="Models\PropertyGroup.cs" />
@@ -601,10 +592,8 @@
     <Compile Include="Persistence\Dtos\UserGroup2NodePermissionDto.cs" />
     <Compile Include="Persistence\Dtos\UserGroupDto.cs" />
     <Compile Include="Persistence\Dtos\UserStartNodeDto.cs" />
-    <Compile Include="Models\UmbracoDomain.cs" />
     <Compile Include="Models\UmbracoObjectTypes.cs" />
     <Compile Include="Models\UserExtensions.cs" />
-    <Compile Include="Models\Validation\RequiredForPersistenceAttribute.cs" />
     <Compile Include="NameValueCollectionExtensions.cs" />
     <Compile Include="ObjectJsonExtensions.cs" />
     <Compile Include="Packaging\ConflictingPackageData.cs" />
@@ -999,11 +988,9 @@
     <Compile Include="Services\IMediaService.cs" />
     <Compile Include="Services\IMediaTypeService.cs" />
     <Compile Include="Services\IMemberService.cs" />
-    <Compile Include="Services\IMembershipUserService.cs" />
     <Compile Include="Services\IMemberTypeService.cs" />
     <Compile Include="Services\Implement\KeyValueService.cs" />
     <Compile Include="Services\IPublicAccessService.cs" />
-    <Compile Include="Services\IUserService.cs" />
     <Compile Include="Services\Implement\LocalizationService.cs" />
     <Compile Include="Services\Implement\LocalizedTextService.cs" />
     <Compile Include="Services\Implement\LocalizedTextServiceFileSources.cs" />
@@ -1062,8 +1049,6 @@
       <Name>Umbraco.Abstractions</Name>
     </ProjectReference>
   </ItemGroup>
-  <ItemGroup>
-    <Folder Include="Models\ContentEditing\" />
-  </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
